Refact reducers
