+++ src/App.jsx
@@ -6,7 +6,7 @@ import Layout from './components/Layout/Layout';
 import VideoView from './pages/VideoView';
 import UserView from './pages/UserView';
 import NotFound from './pages/NotFound';
-import { getUerData } from './store/reducers/Auth.reducer';
+import { getUerData } from './store/reducers/Auth.actionCreator';
 import VideoResults from './pages/VideoResults';
 import RecoverPassword from './components/RecoverPassword';
