+++ src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from 'react';
 import { useDispatch } from 'react-redux';
 import VideoGrid from '../components/VideoGrid';
-import { fetchLabels } from '../store/reducers/Layout.reducer';
+import { fetchLabels } from '../store/reducers/Layout.actionCreator';
 import { fetchAllVideos } from '../store/reducers/Video.actionCreators';
 
 function Home() {
