+++ src/store/reducers/Auth.actions.js
@@ -1,0 +1,4 @@
+export const AUTH_SUCCESS = 'AUTH_SUCCESS';
+export const AUTH_ERROR = 'AUTH_ERROR';
+export const LOGOUT = 'LOGOUT';
+export const USER_SUCCESS = 'USER_SUCCESS';
