+++ src/store/reducers/ChangePassword.actions.js
@@ -1,0 +1,2 @@
+export const SHOW_CHANGE_PASSWORD_FORM = 'SHOW_CHANGE_PASSWORD_FORM';
+export const HIDDE_CHANGE_PASSWORD_FORM = 'HIDDE_CHANGE_PASSWORD_FORM';
