+++ src/store/reducers/ChangePassword.reducer.js
@@ -1,12 +1,7 @@
-const SHOW_CHANGE_PASSWORD_FORM = 'SHOW_CHANGE_PASSWORD_FORM';
-const HIDDE_CHANGE_PASSWORD_FORM = 'HIDDE_CHANGE_PASSWORD_FORM';
-
-const actionBody = (type, payload = null) => ({ type, payload });
-
-export const showChangePasswordForm = () =>
-  actionBody(SHOW_CHANGE_PASSWORD_FORM);
-export const hiddeChangePasswordForm = () =>
-  actionBody(HIDDE_CHANGE_PASSWORD_FORM);
+import {
+  SHOW_CHANGE_PASSWORD_FORM,
+  HIDDE_CHANGE_PASSWORD_FORM,
+} from './ChangePassword.actions';
 
 const initialState = {
   showingChangePasswordForm: false,
