+++ src/store/reducers/Layout.reducer.js
@@ -1,11 +1,8 @@
-import axios from 'axios';
-
-const ACTIONS = {
-  SIDEBAR__TOGGLE: 'SIDEBAR__TOGLE',
-};
-const GET_LABELS_NAME = 'GET_LABELS_NAME';
-const GET_LABELS_ERROR = 'GET_LABELS_ERROR';
-const url = process.env.REACT_APP_BACKEND_URI;
+import {
+  GET_LABELS_NAME,
+  GET_LABELS_ERROR,
+  SIDEBAR__TOGGLE,
+} from './Layout.actions';
 
 const initialState = {
   showSidebar: true,
@@ -13,30 +10,8 @@ const initialState = {
   error: null,
 };
 
-export const fetchLabels = () => {
-  const paramsObject = {
-    page: 1,
-    limit: 14,
-  };
-  return async (dispatch) => {
-    try {
-      const { data } = await axios.get(`${url}labels`, {
-        params: paramsObject,
-      });
-      dispatch({ type: GET_LABELS_NAME, payload: data.labels });
-    } catch (error) {
-      dispatch({ type: GET_LABELS_ERROR, payload: error });
-    }
-  };
-};
-
-const actionBody = (type, payload) => ({ type, payload });
-
-export const sidebarToggle = (payload) =>
-  actionBody(ACTIONS.SIDEBAR__TOGGLE, payload);
-
 function LayoutReducer(state = initialState, action = null) {
-  if (action.type === ACTIONS.SIDEBAR__TOGGLE) {
+  if (action.type === SIDEBAR__TOGGLE) {
     return {
       ...state,
       showSidebar: !state.showSidebar,
